routing gate

+++ routes/web.php
@@ -31,30 +31,30 @@ Route::get('/test', function () {
 Route::controller(ViewTemplateController::class)->group(function () {
     Route::get('/dashboard', function () {
         return view('dashboard.admin.main-dashboard', ['title' => 'Dashboard | Urban Adventure']);
-    })->name('dashboard');
+    })->name('dashboard')->middleware('auth');
 });
 
 Route::controller(CategoryController::class)->group(function () {
-    Route::get('/dashboard/categories', 'allCategory')->name('manage_category.all');
-    Route::get('/dashboard/category/create', 'createCategory')->name('manage_category.create');
-    Route::post('/dashboard/category/create', 'storeCategory')->name('manage_category.store');
-    Route::get('/dashboard/category/{category:id}', 'detailCategory')->name('manage_category.detail');
-    Route::get('/dashboard/category/{category:id}/update', 'updateCategory')->name('manage_category.update');
-    Route::patch('/dashboard/category/{category:id}', 'patchCategory')->name('manage_category.patch');
-    Route::delete('/dashboard/category/{category:id}/delete', 'deleteCategory')->name('manage_category.delete');
+    Route::get('/dashboard/categories', 'allCategory')->name('manage_category.all')->middleware('can:isAdmin');
+    Route::get('/dashboard/category/create', 'createCategory')->name('manage_category.create')->middleware('can:isAdmin');
+    Route::post('/dashboard/category/create', 'storeCategory')->name('manage_category.store')->middleware('can:isAdmin');
+    Route::get('/dashboard/category/{category:id}', 'detailCategory')->name('manage_category.detail')->middleware('can:isAdmin');
+    Route::get('/dashboard/category/{category:id}/update', 'updateCategory')->name('manage_category.update')->middleware('can:isAdmin');
+    Route::patch('/dashboard/category/{category:id}', 'patchCategory')->name('manage_category.patch')->middleware('can:isAdmin');
+    Route::delete('/dashboard/category/{category:id}/delete', 'deleteCategory')->name('manage_category.delete')->middleware('can:isAdmin');
 });
 Route::controller(OrderController::class)->group(function () {
     Route::get('/dashboard/orders', 'allOrder')->name('manage_order.all');
     Route::get('/dashboard/order/create', 'createOrder')->name('manage_order.create');
 });
 Route::controller(ProductController::class)->group(function () {
-    Route::get('/dashboard/products', 'allProduct')->name('manage_product.all');
-    Route::get('/dashboard/product/create', 'createProduct')->name('manage_product.create');
-    Route::post('/dashboard/product/create', 'storeProduct')->name('manage_product.store');
-    Route::get('/dashboard/product/{product}', 'detailProduct')->name('manage_product.detail');
-    Route::get('/dashboard/product/{product}/update', 'updateProduct')->name('manage_product.update');
-    Route::patch('/dashboard/product/{product}', 'patchProduct')->name('manage_product.patch');
-    Route::delete('/dashboard/product/{product}', 'deleteProduct')->name('manage_product.delete');
+    Route::get('/dashboard/products', 'allProduct')->name('manage_product.all')->middleware('can:isAdmin');
+    Route::get('/dashboard/product/create', 'createProduct')->name('manage_product.create')->middleware('can:isAdmin');
+    Route::post('/dashboard/product/create', 'storeProduct')->name('manage_product.store')->middleware('can:isAdmin');
+    Route::get('/dashboard/product/{product}', 'detailProduct')->name('manage_product.detail')->middleware('can:isAdmin');
+    Route::get('/dashboard/product/{product}/update', 'updateProduct')->name('manage_product.update')->middleware('can:isAdmin');
+    Route::patch('/dashboard/product/{product}', 'patchProduct')->name('manage_product.patch')->middleware('can:isAdmin');
+    Route::delete('/dashboard/product/{product}', 'deleteProduct')->name('manage_product.delete')->middleware('can:isAdmin');
 });
 Route::controller(UserController::class)->group(function () {
     Route::get('/dashboard/users', 'allUser')->name('manage_user.all')->middleware('can:isAdmin');
@@ -63,10 +63,7 @@ Route::controller(UserController::class)->group(function () {
     Route::get('/register', 'register')->name('register');
     Route::post('/register', 'attemptRegister')->name('attempt_register');
     Route::get('/logout', 'logout')->name('logout');
-    Route::get('/dashboard/user/{user}', 'detailUser')->name('manage_user.detail');
-    Route::get('/dashboard/user/{user}/update', 'updateUser')->name('manage_user.update');
-    Route::patch('/dashboard/user/{user}', 'patchUser')->name('manage_user.patch');
-    Route::delete('/dashboard/user/{user}', 'deleteUser')->name('manage_user.delete');
+    Route::delete('/dashboard/user/{user}', 'deleteUser')->name('manage_user.delete')->middleware('can:isAdmin');
 });
 Route::controller(CartController::class)->group(function () {
     Route::get('/dashboard/carts', 'allCart')->name('manage_cart.all');
